Implement integration tests with conduit (#3)
This adds a Conduit service container to the test job and performs registration and some other setup, allowing integration tests to get an initialized `matrix_sdk::Client` to send requests to a live homeserver.

Note: This uses conduit mostly because it's smaller and easier to set up than synapse, but we may run into issues if we need to test against endpoints that conduit has not yet implemented. In that case conduit will have to be replaced with synapse. Crossing that bridge when we get to it.